Changed the profit migration from decimal to real

+++ migrations/20230213140453_profit.sql
@@ -2,7 +2,7 @@
 CREATE TABLE profit (
     id INTEGER PRIMARY KEY NOT NULL,
     user_id INTEGER NOT NULL,
-    profit DECIMAL NOT NULL,
-    risk_to_reward DECIMAL NOT NULL,
+    profit REAL NOT NULL,
+    risk_to_reward REAL NOT NULL,
     creation_date VARCHAR(10) NOT NULL
 );
